+++ MapboxAndroidDemo/src/main/java/com/mapbox/mapboxandroiddemo/examples/dds/CircleLayerClusteringActivity.java
@@ -5,7 +5,6 @@ import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.v4.content.ContextCompat;
 import android.support.v7.app.AppCompatActivity;
-import android.util.Log;
 import android.widget.Toast;
 
 import com.mapbox.mapboxandroiddemo.R;
@@ -25,6 +24,8 @@ import com.mapbox.mapboxsdk.utils.BitmapUtils;
 
 import java.net.MalformedURLException;
 import java.net.URL;
+
+import timber.log.Timber;
 
 import static com.mapbox.mapboxsdk.style.expressions.Expression.all;
 import static com.mapbox.mapboxsdk.style.expressions.Expression.division;
@@ -86,8 +87,11 @@ public class CircleLayerClusteringActivity extends AppCompatActivity {
               12.099, -79.045), 3));
 
             addClusteredGeoJsonSource(style);
-            style.addImage("cross-icon-id", BitmapUtils.getBitmapFromDrawable(
-              getResources().getDrawable(R.drawable.ic_cross)));
+            style.addImage(
+              "cross-icon-id",
+              BitmapUtils.getBitmapFromDrawable(getResources().getDrawable(R.drawable.ic_cross)),
+              true
+            );
 
             Toast.makeText(CircleLayerClusteringActivity.this, R.string.zoom_map_in_and_out_instruction,
               Toast.LENGTH_SHORT).show();
@@ -156,7 +160,7 @@ public class CircleLayerClusteringActivity extends AppCompatActivity {
         )
       );
     } catch (MalformedURLException malformedUrlException) {
-      Log.e("dataClusterActivity", "Check the URL " + malformedUrlException.getMessage());
+      Timber.e("Check the URL %s", malformedUrlException.getMessage());
     }